fix: Extract common adapter functionalities
